+++ src/Game/GameObjects/Border.cpp
@@ -12,7 +12,7 @@ void Border::render() const
 {
 	m_pSprite->render(m_position, m_size, m_rotation, m_layer);
 }
-void Border::update(const uint64_t delta)
+void Border::update(const double delta)
 {
 
 }
