+++ src/Game/GameObjects/Border.h
@@ -12,7 +12,7 @@ class Border : public IGameObject {
 public:
 	Border(const std::shared_ptr<RenderEngine::Sprite> pSprite, const glm::vec2 position, const glm::vec2 size, const float rotation, const float layer);
 	virtual void render() const override;
-	virtual void update(const uint64_t delta) override;
+	virtual void update(const double delta) override;
 
 private:
 	std::shared_ptr<RenderEngine::Sprite> m_pSprite;
