+++ src/Game/GameObjects/BrickWall.cpp
@@ -10,7 +10,7 @@ void BrickWall::render() const
 {
 	m_pSprite->render(m_position, m_size, m_rotation, m_layer);
 }
-void BrickWall::update(const uint64_t delta)
+void BrickWall::update(const double delta)
 {
 
 }
