+++ src/Game/GameObjects/IGameObject.h
@@ -6,7 +6,7 @@ class IGameObject {
 public:
 	IGameObject(const glm::vec2 position, const glm::vec2 size, const float rotation, const float layer);
 	virtual void render() const = 0;
-	virtual void update(const uint64_t delta) = 0;
+	virtual void update(const double delta) = 0;
 	virtual ~IGameObject();
 
 protected:
